Add vibranium weapons and tools(they are better that adamantium's things,all the vibranium tools need use obsidian stick to craft) Add vibranium anvil Add vibranium armors' and vibranium anvil's achievements Add vibranium furnace(max heat level is 5) Add vibranium workbench Add vibranium's achievements Make coal block can burn into diamond chunk(4 diamond chunks can make a diamond) in the heat level of 5(use blaze coal) The better the fuel in the furnace be,the faster the items burn Decreased the damage and health amount of most mobs Add a new variant of earth elemental,Wood Elemental,it only can be attacked by axe,battle axe,lava and fire Manure can be used as bonemeal to heal blighted crops Now if someone destroyed others' strong box the server will tell all the players Creeper,Spider,Zombie,Skeleton's health and damage amount will increase as the day count increase(Skeleton will increase its knock back distance) Increase zombie pigman's damage and heal amount Increase Enchantment Looting,Fortune,Protection,Thorns's max level Now the enchantment sharpness,undead killer and Amputation killer can be compatible with each other in vibranium weapons Vibranium weapons and tools can use 15800 xp to enchant at most(use 4 layers of bookshelves and diamond enchantment table),and it can have 4 enchantments at most Blood moon will occurs per 16 days after the blue moon Blue moon only occurs at 128 day

+++ src/net/xiaoyu233/mitemod/miteite/liar/gui/GuiPlayerOverlayLiar.java
@@ -24,7 +24,7 @@ public class GuiPlayerOverlayLiar {
 
         int var4 = MathHelper.f(this.g.h.aN());
         int var5 = MathHelper.f(this.g.h.ax);
-        this.f.setSeed((long)(this.i * 312871));
+        this.f.setSeed(this.i * 312871);
         FoodMetaData var7 = this.g.h.bI();
         int var8 = var7.getNutrition();
 //        AttributeInstance var10 = this.g.h.a(GenericAttributes.a);
@@ -196,7 +196,7 @@ public class GuiPlayerOverlayLiar {
         this.g.C.c("air");
         if (this.g.h.a(Material.h)) {
             int var23 = this.g.h.al();
-            int var35 = MathHelper.f((double)(var23 - 2) * 10.0D / 300.0D);
+                int var35 = MathHelper.f((double)(var23 - 2) * 10.0D / 300.0D);
             int var25 = MathHelper.f((double)var23 * 10.0D / 300.0D) - var35;
 
             for(int var26 = 0; var26 < var35 + var25; ++var26) {
